sysfolk/core/serializer.py
from rest_framework import serializers
from sysfolk.core.models import Person, User


class PersonSerializer(serializers.ModelSerializer):

    class Meta:
        model = Person
        fields = '__all__'

class UserSerializer(serializers.ModelSerializer):

    class Meta:
        model = User
        fields = '__all__'