requirements.txt
dj-database-url==0.4.2
dj-static==0.0.6
Django==1.10.5
django-debug-toolbar==1.6
django-extensions==1.7.6
django-test-without-migrations==0.4
gunicorn==19.6.0
psycopg2==2.6.2
python-decouple==3.0
six==1.10.0
sqlparse==0.2.2
static3==0.7.0
django-jet==1.0.4
google-api-python-client==1.6.2
feedparser==5.2.1

django-cors-headers==2.0.2
djangorestframework==3.5.4
django-rest-auth==0.9.0

## django-filter-1.0.1